Fixed register code, should fully work now
I've just had a meeting with Abu and got help fixing the register code. It turns out that due to security constraints, yo can't use the name "password" in your forms (but "pass" works).

The email errors were due to a missing function redirect. When the email is received it's meant to take you straight to verifyUser() when you click the link in the email, but it runs all the code instead because it wasn't checking for what function was being asked for in the link and barely any of the code was in functions.

So to recap:
- RegisterPersonalDetails is fixed and done
- Bugs in email verification have been fixed for you @liv(edited)
The login page needs sorting out still though, it doesn't seem to post any of the data from the form to the server. It's probably the same issue - renaming it from password to pass and changing the field type to text for invisiform might fix that